Break up utils/primitives.mjs as-is for cleaner update diff.
