+++ src/foundry/common/primitives/module.mjs.d.ts
@@ -1,0 +1,7 @@
+import './array.mjs';
+import './date.mjs';
+import './math.mjs';
+import './number.mjs';
+import './regex.mjs';
+import './set.mjs';
+import './string.mjs';
